+++ oatie/Cargo.toml
@@ -24,6 +24,7 @@ rental = "*"
 wasm-bindgen = "0.2"
 smallvec = "*"
 wasm-typescript-definition = "0.1"
+enumset = "0.3"
 
 [dev-dependencies]
 env_logger = "0.3"
